Extracted ContentBlock, GroupBlock, and Formatted Carousel into separate components from DefaultPage
Components are stored in "ui" folder inside "components" folder.
