+++ src/components/Page/DefaultPage.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import classnames from 'classnames'
 import PropTypes from 'prop-types'
 import {
   useQuery
@@ -16,74 +17,8 @@ import SEO from '../../seo'
 import PixelManager from '../PixelManager'
 
 import { Accordion, Carousel, Row, Loader, Block, Media, Button } from '@christfellowshipchurch/web-ui-kit'
-
-const ContentBlock = ({
-  contentLayout,
-  images,
-  videos,
-  imageAlt,
-  videoUrl,
-  imageRatio,
-  subtitle,
-  title,
-  htmlContent,
-  callToAction,
-  secondaryCallToAction
-}) => (lowerCase(contentLayout) === 'background'
-  ? (
-    <Media
-      imageUrl={get(images, '[0].sources[0].uri', '')}
-      imageAlt={imageAlt}
-      videoUrl={get(videos, '[0].sources[0].uri', '')}
-      ratio={imageRatio}
-    >
-      <Block contentLayout="default" className="text-light">
-        <Block.Title className={`titleWeight`}>
-          {title}
-        </Block.Title>
-
-        <Block.Subtitle className={`font-weight-bold`}>
-          {subtitle}
-        </Block.Subtitle>
-
-        <Block.Body className={`pt-1 pb-4 font-weight-light`}>
-          {htmlContent}
-        </Block.Body>
-
-        <ButtonRow callToAction={callToAction} secondaryCallToAction={secondaryCallToAction} />
-      </Block>
-    </Media>
-  )
-  : (
-    <div className="container py-5">
-      <div className="row">
-        <Block
-          layout={lowerCase(contentLayout)}
-          imageUrl={get(images, '[0].sources[0].uri', '')}
-          imageAlt={imageAlt}
-          videoUrl={videoUrl}
-          ratio={imageRatio}
-        >
-
-          <Block.Subtitle className={`text-muted font-weight-bold`}>
-            {subtitle}
-          </Block.Subtitle>
-
-          <Block.Title className={`titleWeight`}>
-            {title}
-          </Block.Title>
-
-          <Block.Body className={`pt-1 pb-4 font-weight-light`}>
-            {htmlContent}
-          </Block.Body>
-
-          <ButtonRow callToAction={callToAction} secondaryCallToAction={secondaryCallToAction} />
-        </Block>
-      </div>
-    </div>
-  )
-  )
-
+import ContentBlock from '../ui/ContentBlock'
+import GroupBlock from '../ui/GroupBlock'
 
 
 const DefaultPage = ({ title, match: { params: { page } } }) => {
@@ -145,111 +80,6 @@ const DefaultPage = ({ title, match: { params: { page } } }) => {
     </>
   )
 }
-
-const GroupBlock = ({ id, groupLayout }) => {
-  const { loading, error, data } = useQuery(getGroupContentItems, { variables: { id } })
-
-  if (loading) return (
-    <div className="vh-100 vw-100 d-flex justify-content-center align-items-center bg-light">
-      <Loader />
-    </div>
-  )
-  if (error) {
-    console.error("ERROR: ", error)
-    return <h1 className="text-center">There was an error loading block. Please try again.</h1>
-  }
-
-  console.log({ data })
-
-  const groupTitle = data.node.title
-  const groupBody = data.node.htmlContent
-  const blockItems = mapEdgesToNodes(data.node.childContentItemsConnection)
-
-  console.log({ blockItems })
-
-  if (!blockItems || !blockItems.length) return null
-
-  switch (lowerCase(groupLayout)) {
-    case 'row':
-      console.log('Row items:', { blockItems })
-      return (
-        <Row>
-          {blockItems.map((n, i) => n.__typename === 'WebsiteBlockItem'
-            ? <ContentBlock {...n} key={i} /> : null)}
-        </Row>
-      )
-    case 'accordion':
-      return null
-      return (
-        <Accordion
-          blockTitle={groupTitle}
-          blockBody={groupBody}
-        >
-          {blockItems.map((accordionItem, j) => {
-            return (
-              <div key={j} title={accordionItem.title}>
-                <h2>{accordionItem.title}</h2>
-                {accordionItem.htmlContent}
-              </div>
-            )
-          }
-          )}
-        </Accordion>
-      )
-    case 'carousel':
-      // return null
-      console.log('Carousel items:', { blockItems })
-      return (
-        <FormattedCarousel>
-          {blockItems}
-        </FormattedCarousel>
-      )
-    default:
-      return null
-  }
-}
-
-const FormattedCarousel = ({ children }) => (
-  <div className="container py-5">
-    <div className="row justify-content-center">
-      <div className="col-10 col-md-4">
-        <Carousel>
-          {children.map(({
-            __typename,
-            images,
-            imageAlt,
-            videos,
-            videoUrl,
-            subtitle,
-            title,
-            htmlContent,
-            callToAction,
-            secondaryCallToAction
-          }, i) => __typename === 'WebsiteBlockItem'
-              ? (
-                <Block
-                  key={i}
-                  layout={'default'}
-                  media={{
-                    imageUrl: get(images, '[0].sources[0].uri', ''),
-                    imageAlt,
-                    videoUrl: get(videos, '[0].sources[0].uri', ''),
-                    ratio: '1by1',
-                    circle: true
-                  }} >
-                  <Block.Title>{title}</Block.Title>
-                  <Block.Subtitle>{subtitle}</Block.Subtitle>
-                  <Block.Body>{htmlContent}</Block.Body>
-
-                  <ButtonRow callToAction={callToAction} secondaryCallToAction={secondaryCallToAction} />
-                </Block>
-              ) : null
-          )}
-        </Carousel>
-      </div>
-    </div>
-  </div>
-)
 
 const ButtonRow = ({ callToAction, secondaryCallToAction }) => (
   <div>
