+++ src/components/ui/ContentBlock/index.js
@@ -1,0 +1,3 @@
+import ContentBlock from './ContentBlock'
+
+export default ContentBlock
