+++ src/components/ui/FormattedCarousel/index.js
@@ -1,0 +1,3 @@
+import FormattedCarousel from './FormattedCarousel'
+
+export default FormattedCarousel
