+++ src/components/ui/GroupBlock/index.js
@@ -1,0 +1,3 @@
+import GroupBlock from './GroupBlock'
+
+export default GroupBlock
